==> 00_Gathering_Lyrics.ipynb <==
# Gathering "The American Songbook" Lyrics

Since there is no dataset with "The Great American Songbook" lyrics I will need to search the weba dn scrape them from lyrics websites. I chose 'lyricsfreak.com' as my main source for lyrics.

import pandas as pd
import requests
from bs4 import BeautifulSoup
import time

## Initial html research

Learning how this html page works and how I can grab the hyperlinks.

url = 'https://www.lyricsfreak.com/c/cole+porter/'
res = requests.get(url)

res.status_code

soup = BeautifulSoup(res.content, 'lxml')

soup.find('a', {'class': 'lf-link lf-link--primary'})

cole_porter_table = soup.find('div', {'class': 'lf-list js-sort-table'})

cole_porter_songs = []

for row in cole_porter_table.find_all('a'):
    cp_song = {
        'song_title': row.attrs['title'],
        'href': row.attrs['href']
    }
    cole_porter_songs.append(cp_song)
# pd.DataFrame(cole_porter_songs)
cole_porter_songs[:5]

song_url = 'https://www.lyricsfreak.com/c/cole+porter/all+of+you_20816177.html'
song_res = requests.get(song_url)

song_res.status_code

song_soup = BeautifulSoup(song_res.content, 'lxml')

song_soup.find('div', {'class': 'lyrictxt js-lyrics js-share-text-content'})

song_soup.find('div', {'class': 'lyrictxt js-lyrics js-share-text-content'}).attrs

song_soup.find('div', {'class': 'lyrictxt js-lyrics js-share-text-content'}).attrs['data-container-id']

song_soup.find('div', {'class': 'lyrictxt js-lyrics js-share-text-content'}).text

## Grabbing Cole Porter Lyrics

This is a test run to grab Cole Porter lyrics. This will be the basis for my function. 

# This code has been taken from the web scraping lab solution code in week 5
# Start with an empty list
cole_porter_lyrics = []

# Loop through each lyric 
for lyric in cole_porter_songs:
    print('Scraping {}'.format(lyric['song_title']))
    
    href = lyric['href']
    lyric_url = f'https://www.lyricsfreak.com/c/cole+porter/{href}'
    
    # Use requests library to get the content from each lyric page
    lyric_res = requests.get(lyric_url)
    
    # Create soup object from lyrics html
    lyric_soup = BeautifulSoup(lyric_res.content, 'lxml')
    lyrics = {}
    
    # Songwriter 
    lyrics['song_writer'] = 'Cole Porter'
    
    # Song name
    lyrics['song_name'] = lyric['song_title']
    
    # Creating a lyrics key
    lyrics['lyrics'] = lyric_soup.find('div', {'class': 'lyrictxt js-lyrics js-share-text-content'}).text
    
    # Song link
    lyrics['song_link'] = lyric_url
    cole_porter_lyrics.append(lyrics)
        
    # Putting a sleep timer on the scraping
    time.sleep(20)

cole_porter_df = pd.DataFrame(cole_porter_lyrics)

cole_porter_df.head()

## Creating Functions to grab lyrics

The three functions below are to make the process of gathering lyrics a little easier and with a little less effort.

# This function is to grab the list of songs and hyperlinks
def gathering_songs(url, song_writer):
    
    songwriter_url = url
    
    res_songwriter = requests.get(songwriter_url)
    
    soup_songwriter = BeautifulSoup(res_songwriter.content, 'lxml')
    
    songwriter_table = soup_songwriter.find('div', {'class': 'lf-list js-sort-table'})
    
    songwriter_songs = []
    
    for row in songwriter_table.find_all('a'):
        try:
            songs = {
                'song_writer': song_writer,
                'song_title': row.attrs['title'],
                'href': row.attrs['href']
            }
            songwriter_songs.append(songs)
        except:
            pass
    return songwriter_songs

# This function is to gather all the lyrics from the list of songs
def gathering_lyrics(song_list, url):
    songwriter_lyrics = []

    # Loop through each lyric 
    for lyric in song_list:
        print('Scraping {}'.format(lyric['song_title']))

        href = lyric['href']
        lyric_url = f'{url} + {href}'

        # Use requests library to get the content from each lyric page
        lyric_res = requests.get(lyric_url)

        # Create soup object from lyrics html
        lyric_soup = BeautifulSoup(lyric_res.content, 'lxml')
        lyrics = {}

        # Songwriter 
        lyrics['song_writer'] = lyric['song_writer']

        # Song name
        lyrics['song_name'] = lyric['song_title']

        # Creating a lyrics key
        lyrics['lyrics'] = lyric_soup.find('div', {'class': 'lyrictxt js-lyrics js-share-text-content'}).text

        # Song link
        lyrics['song_link'] = lyric_url
        songwriter_lyrics.append(lyrics)

        # Putting a sleep timer on the scraping
        time.sleep(10)
    return songwriter_lyrics

# This function is combining the two functions above into one succinct function.
def grabbing_songs_and_lyrics(url, songwriter):
    songs_list = gathering_songs(url, songwriter)
    
    print("Songs have been gathered")
    print("Now to grab lyrics")
    
    all_lyrics =  gathering_lyrics(songs_list, url)
    
    return all_lyrics

## Grabbing all the lyrics!!

#### Oscar Hammerstein

oscar_hammerstein_songs = gathering_songs('https://www.lyricsfreak.com/o/oscar+hammerstein/', "Oscar Hammerstein")

oscar_hammerstein = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/o/oscar+hammerstein/', "Oscar Hammerstein")

oscar_hammerstein_df = pd.DataFrame(oscar_hammerstein)

#### Hoagy Carmichael

hoagy_carmichael = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/h/hoagy+carmichael/', 'Hoagy Carmichael')

hoagy_carmichael_df = pd.DataFrame(hoagy_carmichael)

hoagy_carmichael_df.head()

#### Harold Arlen

harold_arlen = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/h/harold+arlen/', 'Harold Arlen')

harold_arlen_df = pd.DataFrame(harold_arlen)

harold_arlen_df.head()

### Concating some of the lyrics into a CSV

some_composer_df = pd.concat([harold_arlen_df,
           hoagy_carmichael_df,
           oscar_hammerstein_df,
           cole_porter_df], axis = 0)

some_composer_df.to_csv('datasets/arlen_carmichael_hammerstein_porter.csv', index=False)

#### Fats Waller

fats_waller = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/f/fats+waller/', 'Fats Waller')

fats_waller_df = pd.DataFrame(fats_waller)

fats_waller_df.shape

fats_waller_df['lyrics'][0]

#### George Gershwin

The George Gershwin list of lyrics were set up differently. So I had to create slightly different code to work with the different errors that were popping up. There were songs on the webpage that didn't have a hyperlink and it was throwing an error in my code. I resolved it, by skipping those empty lyrics. 

# https://www.lyricsfreak.com/g/george+gershwin/
url = 'https://www.lyricsfreak.com/g/george+gershwin/'
res = requests.get(url)
soup = BeautifulSoup(res.content, 'lxml')
george_gershwin_table = soup.find('div', {'class': 'lf-list js-sort-table'})

george_gershwin_songs = []

for row in george_gershwin_table.find_all('a'):
    try:
        gg_song = {
            'song_title': row.attrs['title'],
            'href': row.attrs['href']             
        }
        george_gershwin_songs.append(gg_song)
    except:
        pass
#     george_gershwin_songs.append(gg_song)
# pd.DataFrame(cole_porter_songs)
george_gershwin_songs

george_gershwin = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/g/george+gershwin/', 'George Gershwin')

george_gershwin_df = pd.DataFrame(george_gershwin)

george_gershwin_df

#### Irving Berlin

# https://www.lyricsfreak.com/i/irving+berlin/
irving_berlin = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/i/irving+berlin/', 'Irving Berlin')

irving_berlin_df = pd.DataFrame(irving_berlin)

irving_berlin_df

#### Johnny Mercer

# https://www.lyricsfreak.com/j/johnny+mercer/
johnny_mercer = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/j/johnny+mercer/', 'Johnny Mercer')

johnny_mercer_df = pd.DataFrame(johnny_mercer)

#### Jerome Kern

# https://www.lyricsfreak.com/j/jerome+kern/
jerome_kern = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/j/jerome+kern/', 'Jerome Kern')

jerome_kern_df = pd.DataFrame(jerome_kern)

jerome_kern_df

#### Richard Rodgers

richard_rodgers = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/r/richard+rodgers/', 'Richard Rodgers')

richard_rodgers_df = pd.DataFrame(richard_rodgers)

### Concating all the second round of composers onto a CSV

more_composer_df = pd.concat([fats_waller_df,
                              george_gershwin_df,
                              irving_berlin_df,
                              johnny_mercer_df,
                              jerome_kern_df,
                              richard_rodgers_df], axis = 0)

more_composer_df.to_csv('datasets/berlin_gershwin_kern_mercer_rodgers_waller.csv', index=False)

# Scraping Disney Renaissance songs

#### Little Mermaid

little_mermaid = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/t/the+little+mermaid/', 'Little Mermaid')

little_mermaid_df = pd.DataFrame(little_mermaid)

#### Beauty and the Beast

beauty_beast = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/b/beauty+and+the+beast/', 'Beauty and the Beast')

beauty_beast_df = pd.DataFrame(beauty_beast)

#### Aladdin

aladdin = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/a/aladdin/', 'Aladdin')

aladdin_df = pd.DataFrame(aladdin)

#### The Lion King

lion_king = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/l/lion+king/', 'Lion King')

lion_king_df = pd.DataFrame(lion_king)

#### Pocahontas

pocahontas = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/p/pocahontas/', 'Pocahontas')

pocahontas_df = pd.DataFrame(pocahontas)

#### Hercules

hercules = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/d/disneys+hercules/', 'Hercules')

hercules_df = pd.DataFrame(hercules)

#### Mulan

mulan = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/m/mulan/', 'Mulan')

mulan_df = pd.DataFrame(mulan)

#### Tarzan

tarzan = grabbing_songs_and_lyrics('https://www.lyricsfreak.com/d/disneys+tarzan/', 'Tarzan')

tarzan_df = pd.DataFrame(tarzan)

## Concating Disney Renaissance songs to CSV

disney_renaissance = pd.concat([tarzan_df,
           mulan_df,
           hercules_df,
           pocahontas_df,
           lion_king_df,
           aladdin_df,
           beauty_beast_df,
           little_mermaid_df], axis=0)

disney_renaissance.to_csv('datasets/disney_renaissance.csv', index=False)
