==> 01_Gathering_Disney_Lyrics.ipynb <==
# Gathering all the Disney Lyrics

Since I was scraping lyrics from a different website I need to make a slight adjustment to my code. Other than that this notebook is just scraping for all the Disney lyrics I used for the project.

import pandas as pd
import requests
from bs4 import BeautifulSoup
import time
import re

def gathering_songs(url, song_writer):
    
    songwriter_url = url
    
    res_songwriter = requests.get(songwriter_url)
    
    soup_songwriter = BeautifulSoup(res_songwriter.content, 'lxml')
    
    songwriter_table = soup_songwriter.find('table', {'class': 'songs'})
    
    songwriter_songs = []
    
    for row in songwriter_table.find_all('a'):
        try:
            songs = {
                'song_writer': song_writer,
                'song_title': row.text,
                'href': row.attrs['href']
            }
            songwriter_songs.append(songs)
        except:
            pass
    return songwriter_songs

# This code has been taken from the web scraping lab in week 5
# Start with an empty list
def gathering_lyrics(song_list):
    songwriter_lyrics = []

    # Loop through each lyric 
    for lyric in song_list:
        print('Scraping {}'.format(lyric['song_title']))

        href = lyric['href']
        lyric_url = f'http://www.disneyclips.com/lyrics/{href}'

        # Use requests library to get the content from each lyric page
        lyric_res = requests.get(lyric_url)

        # Create soup object from lyrics html
        lyric_soup = BeautifulSoup(lyric_res.content, 'lxml')
        lyrics = {}

        # Songwriter 
        lyrics['song_writer'] = lyric['song_writer']

        # Song name
        lyrics['song_name'] = lyric['song_title']

        # Creating a lyrics key
#         lyrics['lyrics'] = lyric_soup.find('div', {'class': 'main'})('p')[1].text
        lyrics['lyrics'] = str(lyric_soup.find_all('p')[1]).\
        replace('<br/>', ' ').\
        replace('<p>', '').\
        replace('</p>', '').\
        replace('<i>Lyrics transcribed by Disneyclips.com</i>', '').\
        replace('  ', ' ').rstrip()
        # Song link
        lyrics['song_link'] = lyric_url
        songwriter_lyrics.append(lyrics)

        # Putting a sleep timer on the scraping
        time.sleep(5)
    return songwriter_lyrics

def grabbing_songs_and_lyrics(url, songwriter):
    songs_list = gathering_songs(url, songwriter)
    
    print(f"{songwriter} list of songs have been gathered")
    print(f"Now grabbing {songwriter} song lyrics")
    
    all_lyrics =  gathering_lyrics(songs_list)
    print()
    print('='*50)
    print()
    return all_lyrics

disney_songs_dict = {
    '101 Dalmations': 'http://www.disneyclips.com/lyrics/101-dalmatians.html',
    'Alice In Wonderland': 'http://www.disneyclips.com/lyrics/alice-in-wonderland.html',
    'The Aristocats': 'http://www.disneyclips.com/lyrics/the-aristocats.html',
    'Cinderella': 'http://www.disneyclips.com/lyrics/cinderella.html',
    'Dumbo': 'http://www.disneyclips.com/lyrics/dumbo.html',
    'Frozen': 'http://www.disneyclips.com/lyrics/frozen.html',
    'The Hunchback of Notre Dame': 'http://www.disneyclips.com/lyrics/the-hunchback-of-notre-dame.html',
    'The Jungle Book': 'http://www.disneyclips.com/lyrics/the-jungle-book.html',
    'Lady and the Tramp': 'http://www.disneyclips.com/lyrics/lady-and-the-tramp.html',
    'Mary Poppins': 'http://www.disneyclips.com/lyrics/mary-poppins.html',
    'Peter Pan': 'http://www.disneyclips.com/lyrics/peter-pan.html',
    'Pinocchio': 'http://www.disneyclips.com/lyrics/pinocchio.html',
    'The Princess and the Frog': 'http://www.disneyclips.com/lyrics/the-princess-and-the-frog.html',
    'Robin Hood': 'http://www.disneyclips.com/lyrics/robin-hood.html',
    'Sleeping Beauty': 'http://www.disneyclips.com/lyrics/sleeping-beauty.html',
    'Snow White': 'http://www.disneyclips.com/lyrics/snow-white-and-the-seven-dwarfs.html',
    'Coco': 'http://www.disneyclips.com/lyrics/coco.html',
    'Bedknobs and Broomsticks': 'http://www.disneyclips.com/lyrics/bedknobs-and-broomsticks.html',
    'Bambi': 'https://www.disneyclips.com/lyrics/bambi.html',
    'The Great Mouse Dectective': 'https://www.disneyclips.com/lyrics/the-great-mouse-detective.html',
    'Moana': 'https://www.disneyclips.com/lyrics/moana.html',
    'The Nightmare Before Christmas': 'https://www.disneyclips.com/lyrics/the-nightmare-before-christmas.html',
    'Enchanted': 'https://www.disneyclips.com/lyrics/enchanted.html',
    'Tangled': 'https://www.disneyclips.com/lyrics/tangled.html',
    'The Fox and the Hound': 'https://www.disneyclips.com/lyrics/the-fox-and-the-hound.html',
    'The Emperor\'s New Groove': 'http://www.disneyclips.com/lyrics/the-emperors-new-groove.html'
}

all_songs = []
for songs, site in disney_songs_dict.items():
#     print(href, songs)
    
    grabbed_songs = grabbing_songs_and_lyrics(songwriter = songs, url = site)
#     print(grabbed_songs)
    all_songs.append(grabbed_songs)
#     print(all_songs)
# pd.DataFrame(all_songs)

all_songs_df = []
for i, _ in enumerate(all_songs):

    df = pd.DataFrame(all_songs[i])
    all_songs_df.append(df)

all_songs_df = pd.concat(all_songs_df)

all_songs_df

all_songs_df.to_csv('datasets/disney_songs.csv', index=False)

more_disney_songs = {
    'Brave': 'http://www.disneyclips.com/lyrics/brave.html',
    'Brother Bear': 'http://www.disneyclips.com/lyrics/brother-bear.html',
    'Frozen 2': 'http://www.disneyclips.com/lyrics/frozen2.html',
    'A Goofy Movie': 'http://www.disneyclips.com/lyrics/a-goofy-movie.html',
    'High School Musical': 'http://www.disneyclips.com/lyrics/highschoolmusical.html',
    'Oliver and Company': 'http://www.disneyclips.com/lyrics/oliver-and-company.html',
    'The Rescuers': 'http://www.disneyclips.com/lyrics/the-rescuers.html',
    'The Sword and the Stone': 'http://www.disneyclips.com/lyrics/the-sword-in-the-stone.html',
    'Winnie the Pooh': 'http://www.disneyclips.com/lyrics/winnie-the-pooh.html'
}

disney_songs = []
for songs, site in more_disney_songs.items():
#     print(href, songs)
    
    grabbed_songs = grabbing_songs_and_lyrics(songwriter = songs, url = site)
#     print(grabbed_songs)
    disney_songs.append(grabbed_songs)
#     print(all_songs)
# pd.DataFrame(all_songs)

disney_songs_df = []
for i, _ in enumerate(disney_songs):

    df = pd.DataFrame(disney_songs[i])
    disney_songs_df.append(df)

disney_songs_df = pd.concat(disney_songs_df)

disney_songs_df

disney_songs_df.to_csv('datasets/more_disney_songs.csv', index=False)
