==> 02_Combining_Songs.ipynb <==
In this notebook I combine all the songs that I scraped from the two lyric websites into one CSV for easier access.

import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import regex as re
from nltk.corpus import stopwords

## First combining "The Great American Songbook" lyrics

songs = pd.read_csv('datasets/arlen_carmichael_hammerstein_porter.csv')

more_songs = pd.read_csv('datasets/berlin_gershwin_kern_mercer_rodgers_waller.csv')

american_songbook = pd.concat([songs, more_songs], axis=0)

american_songbook.head()

american_songbook.song_writer.value_counts()

## Labeling the lyrics with the appropriate classification ID before combining with Disney.

american_songbook['american_songbook'] = 1

american_songbook['disney_renaissance'] = 0

american_songbook.head()

## Combining Disney Lyrics

disney_renaissance = pd.read_csv('datasets/disney_renaissance.csv')

Again labeling the the data with the classification ID before combining.

disney_renaissance['disney_renaissance'] = 1

disney_renaissance['american_songbook'] = 0

disney_renaissance.head()

disney = pd.read_csv('datasets/disney_songs.csv')

more_disney = pd.read_csv('datasets/more_disney_songs.csv')

disney_songs = pd.concat([disney, more_disney], axis=0)

disney_songs.head()

disney_songs['disney_renaissance'] = 0

disney_songs['american_songbook'] = 0

disney_songbook = pd.concat([disney_songs, disney_renaissance], axis=0)

disney_songbook.head()

disney_songbook['song_writer'].value_counts().sort_index()

## Combining all the lyrics together!!

all_lyrics = pd.concat([american_songbook,
           disney_songbook])

# Checking to see if the songbooks are as balanced as possible
all_lyrics['american_songbook'].value_counts(normalize=True)

## Figuring out the code to remove unnessary characters

all_lyrics.iloc[0]

all_lyrics.iloc[0]['lyrics']

example1 = BeautifulSoup(all_lyrics.iloc[0]['lyrics'])

# Print the raw review and then the output of get_text(), for comparison
print(example1.get_text())

# Use regular expressions to do a find-and-replace
letters_only = re.sub("[^a-zA-Z'-]",           
                      " ",                  
                      example1.get_text().rstrip())

letters_only

songs['lyrics'][0] = letters_only

songs['lyrics'][0]

lower_case = letters_only.lower()

lower_case

words = lower_case.split()

words[:10]

all_lyrics = all_lyrics.sort_values('song_writer')

all_lyrics.to_csv('datasets/all_lyrics.csv', index=False)

all_lyrics = pd.read_csv('datasets/all_lyrics.csv')

all_lyrics

## Cleaning up the lyrics

# This function has been taken from Matt Brems NLP I lecture. Very little has been changed.
# Only difference is a change in the function name
def lyrics_to_words(raw_text):
    
    text = BeautifulSoup(raw_text).get_text()
    
    letters_only = re.sub("[^a-zA-Z'-]", " ", text)
    
    words = letters_only.lower().split()
    
#     stops = set(stopwords.words('english'))
    
#     meaningful_words = [w for w in words if w not in stops]

    return(" ".join(words))

total_songs = all_lyrics.shape[0]

# This code is heavily borrowed from NLP I lesson by Matt Brems.
# There have been some slight modifications made to fit this project.
# Initialize an empty list to hold the cleaned songs
clean_songs = []

print("Cleaning and parsing the lyrics...")

# Instantiate counter.
j = 0

# Looping through the lyrics...
for songs in all_lyrics['lyrics']:
    
    # Convert lyrics to separate words, then append to clean_songs.
    clean_songs.append(lyrics_to_words(songs))
    
    # If the index is divisible by 50, print a message.
    if (j + 1) % 50 == 0:
        print(f'Lyric {j + 1} of {total_songs}.')
    
    j += 1

# Creating a dataframe for the clean titles
clean_songs_df = pd.DataFrame(clean_songs, columns=['clean_lyrics'])

clean_songs_df

all_lyrics.shape

clean_songs_df.shape

# Removing all characters and stopwords

I was interested to see how differently the model scores would be if I removed the stopwords. 

# This function has been taken from Matt Brems NLP I lecture. Very little has been changed.
# Only difference is a change in the function name
def lyrics_to_words_stopwords(raw_text):
    
    text = BeautifulSoup(raw_text).get_text()
    
    letters_only = re.sub("[^a-zA-Z]", " ", text)
    
    words = letters_only.lower().split()
    
    stops = set(stopwords.words('english'))
    
    meaningful_words = [w for w in words if w not in stops]

    return(" ".join(words))

# This code is heavily borrowed from NLP I lesson by Matt Brems.
# There have been some slight modifications made to fit this project.
# Initialize an empty list to hold the cleaned songs
clean_songs_stopwords = []

print("Cleaning and parsing the lyrics...")

# Instantiate counter.
j = 0

# Looping through the lyrics...
for songs in all_lyrics['lyrics']:
    
    # Convert lyrics to separate words, then append to clean_songs.
    clean_songs_stopwords.append(lyrics_to_words_stopwords(songs))
    
    # If the index is divisible by 50, print a message.
    if (j + 1) % 50 == 0:
        print(f'Lyric {j + 1} of {total_songs}.')
    
    j += 1

# Creating a dataframe for the clean titles
clean_songs_stopwords_df = pd.DataFrame(clean_songs_stopwords, columns=['clean_lyrics_with_stopwords'])

clean_songs_stopwords_df

all_lyrics_clean = pd.concat([all_lyrics, clean_songs_df, clean_songs_stopwords_df], axis=1, join='inner')

all_lyrics_clean.to_csv('datasets/all_lyrics_add_clean_lyrics_col', index=False)
