==> 04_Modeling.ipynb <==
This notebook is all modeling. I was interested in trying out as many different models as possible to see how well it can do to predict between the two lyric songbooks. I also was interested to see how the models would compare between a CountVectorizer and TFIDFVectorizer. All the score are towards the bottom of the notebook. 

# Importing libraries
import pandas as pd

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import GaussianNB
from sklearn.base import TransformerMixin
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier, GradientBoostingClassifier,\
                             AdaBoostClassifier
from sklearn.svm import SVC

songs = pd.read_csv('datasets/cleaned_lyrics_with_sentiment_analysis')

# Setting up my train_test_split

X = songs['clean_lyrics']
y = songs['american_songbook']

X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    test_size=0.25,
                                                    stratify=y,
                                                    random_state=42)

# CountVectorizer With StopWords

pipe = Pipeline([
    ('cvec', CountVectorizer()),
    ('lr', LogisticRegression(max_iter=5000))
])

pipe_params = {
    'cvec__max_features': [2000, 3000, 4000, 5000],
    'cvec__min_df': [1, 2, 3 , 4],
    'cvec__max_df': [.9, .95],
    'cvec__ngram_range': [(1,1), (1,2), (1,3)],
    'cvec__stop_words': [None, 'english'],
}

gs = GridSearchCV(pipe,
                  pipe_params,
                  cv=5)

gs.fit(X_train, y_train)

gs.best_score_

gs.score(X_test, y_test)

gs.best_params_

# TFIDFVectorizer

tfidf_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('lr', LogisticRegression())
])

tfidf_pipe_params = {
    'tfidf__max_features': [2000, 3000, 4000, 5000],
    'tfidf__min_df': [1, 2, 3 , 4],
    'tfidf__max_df': [.9, .95],
    'tfidf__ngram_range': [(1,1), (1,2), (2,2)],
    'tfidf__stop_words': [None, 'english']
}

gs_tfidf = GridSearchCV(tfidf_pipe,
                        tfidf_pipe_params,
                        cv=5)

gs_tfidf.fit(X_train, y_train)

gs_tfidf.best_score_

gs_tfidf.score(X_test, y_test)

gs_tfidf.best_params_

# Extra Trees Classifier

et_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('et', ExtraTreesClassifier(random_state=42))
])

et_pipe_params = {
    'tfidf__max_features': [3000],
    'tfidf__min_df': [1],
    'tfidf__max_df': [.9],
    'tfidf__ngram_range': [(1,1)],
    'tfidf__stop_words': [None],
    'et__n_estimators': [50, 100],
    'et__max_features': [None, 'auto'],
    'et__max_depth': [None, 2, 3, 4]
}

gs_et = GridSearchCV(et_pipe,
                     et_pipe_params,
                     cv=5)

gs_et.fit(X_train, y_train)

gs_et.best_score_

gs_et.score(X_test, y_test)

gs_et.best_params_

### With CountVectorizer

et_cv_pipe = Pipeline([
    ('cv', CountVectorizer()),
    ('et', ExtraTreesClassifier(random_state=42))
])

et_cv_pipe_params = {
    'cv__max_features': [2000],
    'cv__min_df': [2],
    'cv__max_df': [.95],
    'cv__ngram_range': [(1,1)],
    'cv__stop_words': [None],
    'et__n_estimators': [50, 100],
    'et__max_features': [None, 'auto'],
    'et__max_depth': [None, 2, 3, 4]
}

gs_et_cv = GridSearchCV(et_cv_pipe,
                        et_cv_pipe_params,
                        cv=5)

gs_et_cv.fit(X_train, y_train)

gs_et_cv.best_score_

gs_et_cv.score(X_test, y_test)

gs_et_cv.best_params_

# Random Forest Classifier

rf_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('rf', RandomForestClassifier(random_state=42))
])

rf_pipe_params = {
    'tfidf__max_features': [3000],
    'tfidf__min_df': [1],
    'tfidf__max_df': [.9],
    'tfidf__ngram_range': [(1,1)],
    'tfidf__stop_words': [None],
    'rf__n_estimators': [50, 100],
    'rf__max_features': [None, 'auto'],
    'rf__max_depth': [None, 2, 3, 4]
}

gs_rf = GridSearchCV(rf_pipe,
                     rf_pipe_params,
                     cv=5)

gs_rf.fit(X_train, y_train)

gs_rf.best_score_

gs_rf.score(X_test, y_test)

gs_rf.best_params_

### With CountVectorizer

rf_cv_pipe = Pipeline([
    ('cv', CountVectorizer()),
    ('rf', RandomForestClassifier(random_state=42))
])

rf_cv_pipe_params = {
    'cv__max_features': [2000],
    'cv__min_df': [2],
    'cv__max_df': [.95],
    'cv__ngram_range': [(1,1)],
    'cv__stop_words': [None],
    'rf__n_estimators': [50, 100],
    'rf__max_features': [None, 'auto'],
    'rf__max_depth': [None, 2, 3, 4]
}

gs_rf_cv = GridSearchCV(rf_cv_pipe,
                     rf_cv_pipe_params,
                     cv=5)

gs_rf_cv.fit(X_train, y_train)

gs_rf_cv.best_score_

gs_rf_cv.score(X_test, y_test)

gs_rf_cv.best_params_

# SVC

SVC_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('SVC', SVC(random_state=42))
])

SVC_pipe_params = {
    'tfidf__max_features': [3000],
    'tfidf__min_df': [1],
    'tfidf__max_df': [.9],
    'tfidf__ngram_range': [(1,1)],
    'tfidf__stop_words': [None],
    'SVC__C': [1, .1, .01, .001],
    'SVC__degree': [3, 4, 5],
    'SVC__kernel': ['linear'],
    'SVC__probability': [True]
}

gs_SVC = GridSearchCV(SVC_pipe,
                     SVC_pipe_params,
                     cv=5)

gs_SVC.fit(X_train, y_train)

gs_SVC.best_score_

gs_SVC.score(X_test, y_test)

gs_SVC.best_params_

#### With CountVectorizer

SVC_cv_pipe = Pipeline([
    ('cv', CountVectorizer()),
    ('SVC', SVC(random_state=42))
])

SVC_cv_pipe_params = {
    'cv__max_features': [2000],
    'cv__min_df': [3],
    'cv__max_df': [.95],
    'cv__ngram_range': [(1,1)],
    'cv__stop_words': [None],
    'SVC__C': [1],
    'SVC__degree': [3],
    'SVC__kernel': ['linear']
}

gs_SVC_cv = GridSearchCV(SVC_cv_pipe,
                     SVC_cv_pipe_params,
                     cv=5)

gs_SVC_cv.fit(X_train, y_train)

gs_SVC_cv.best_score_

gs_SVC_cv.score(X_test, y_test)

gs_SVC_cv.best_params_

# Guassian Bayes

# adapted from https://stackoverflow.com/questions/28384680/scikit-learns-pipeline-a-sparse-matrix-was-passed-but-dense-data-is-required
class DenseTransformer(TransformerMixin):
    def fit(self, X, y=None, **fit_params):
        return self
    
    def transform(self, X, y=None, **fit_params):
        return X.toarray()

gauss_pipe = Pipeline([
    ('cv', CountVectorizer()),
    ('_', DenseTransformer()),
    ('gauss', GaussianNB())
])

gauss_params = {
    'cv__max_features': [2000, 3000, 4000],
    'cv__min_df': [1, 2, 3],
    'cv__max_df': [.95],
    'cv__ngram_range': [(1,1), (1,2)],
    'cv__stop_words': [None, 'english'],
}

gauss_gs = GridSearchCV(gauss_pipe, gauss_params, verbose=1, n_jobs=-1)

gauss_gs.fit(X_train, y_train)

gauss_gs.best_score_

gauss_gs.score(X_test, y_test)

#### TFIDFVectorizer

gauss_tfidf_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('_', DenseTransformer()),
    ('gauss', GaussianNB())
])

gauss_tfidf_params = {
    'tfidf__max_features': [2000, 3000, 4000],
    'tfidf__min_df': [1, 2, 3],
    'tfidf__max_df': [.95],
    'tfidf__ngram_range': [(1,1), (1,2)],
    'tfidf__stop_words': [None, 'english'],
}

gauss_gs_tfidf = GridSearchCV(gauss_tfidf_pipe, gauss_tfidf_params, verbose=1, n_jobs=-1)

gauss_gs_tfidf.fit(X_train, y_train)

gauss_gs_tfidf.best_score_

gauss_gs_tfidf.score(X_test, y_test)

# AdaBoost

ADA_cv_pipe = Pipeline([
    ('cv', CountVectorizer()),
    ('ADA', AdaBoostClassifier(random_state=42))
])

ADA_cv_pipe_params = {
    'cv__max_features': [2000, 3000, 4000],
    'cv__min_df': [1, 2, 3],
    'cv__max_df': [.95],
    'cv__ngram_range': [(1,1), (1,2)],
    'cv__stop_words': [None, 'english'],
    'ADA__n_estimators': [50, 100]
}

gs_ADA_cv = GridSearchCV(ADA_cv_pipe,
                     ADA_cv_pipe_params,
                     cv=5)

gs_ADA_cv.fit(X_train, y_train)

gs_ADA_cv.best_score_

gs_ADA_cv.score(X_test, y_test)

gs_ADA_cv.best_params_

#### TFIDFVector

ADA_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('ADA', AdaBoostClassifier(random_state=42))
])

ADA_pipe_params = {
    'tfidf__max_features': [2000, 3000, 4000],
    'tfidf__min_df': [1, 2, 3],
    'tfidf__max_df': [.95],
    'tfidf__ngram_range': [(1,1), (1,2)],
    'tfidf__stop_words': [None, 'english'],
    'ADA__n_estimators': [50, 100]
}

gs_ADA = GridSearchCV(ADA_pipe,
                     ADA_pipe_params,
                     cv=5)

gs_ADA.fit(X_train, y_train)

gs_ADA.best_score_

gs_ADA.score(X_test, y_test)

# Gradient Boosting

gb_cv_pipe = Pipeline([
    ('cv', CountVectorizer()),
    ('gb', GradientBoostingClassifier(random_state=42))
])

gb_cv_pipe_params = {
    'cv__max_features': [2000, 3000, 4000],
    'cv__min_df': [1, 2, 3],
    'cv__max_df': [.95],
    'cv__ngram_range': [(1,1), (1,2)],
    'cv__stop_words': [None, 'english'],
    'gb__n_estimators': [50, 100, 200]
}

gs_gb_cv = GridSearchCV(gb_cv_pipe,
                     gb_cv_pipe_params,
                     cv=5)

gs_gb_cv.fit(X_train, y_train)

gs_gb_cv.best_score_

gs_gb_cv.score(X_test, y_test)

gs_gb_cv.best_params_

#### TFIDFVectorizer

gb_pipe = Pipeline([
    ('tfidf', TfiTfidfVectorizerctorizer()),
    ('gb', GradientBoostingClassifier(random_state=42))
])

gb_pipe_params = {
    'tfidf__max_features': [2000, 3000, 4000],
    'tfidf__min_df': [1, 2, 3],
    'tfidf__max_df': [.95],
    'tfidf__ngram_range': [(1,1), (1,2)],
    'tfidf__stop_words': [None, 'english'],
    'gb__n_estimators': [50, 100, 200]
}

gs_gb = GridSearchCV(gb_pipe,
                     gb_pipe_params,
                     cv=5)

gs_gb.fit(X_train, y_train)

gs_gb.best_score_

gs_gb.score(X_test, y_test)

gs_gb.best_params_

# All Scores combined

print('Training Best Score, Logisitic Regression, CVEC:', gs.best_score_)
print('Testing Score, Logisitic Regression, CVEC:', gs.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Logisitic Regression, TFIDF:', gs_tfidf.best_score_)
print('Testing Score, Logisitic Regression, TFIDF:', gs_tfidf.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Extra Tree Classifier, CVEC:', gs_et_cv.best_score_)
print('Testing Score, Extra Tree Classifier, CVEC:', gs_et_cv.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Extra Tree Classifier, TFIDF:', gs_et.best_score_)
print('Testing Score, Extra Tree Classifier, TFIDF:', gs_et.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Random Forest Classifier, CVEC:', gs_rf_cv.best_score_)
print('Testing Score, Random Forest Classifier, CVEC:', gs_rf_cv.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Random Forest Classifier, TFIDF:', gs_rf.best_score_)
print('Testing Score, Random Forest Classifier, TFIDF:', gs_rf.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, SVC, CVEC:', gs_SVC_cv.best_score_)
print('Testing Score, SVC, CVEC:', gs_SVC_cv.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, SVC, TFIDF:', gs_SVC.best_score_)
print('Testing Score, Extra SVC, TFIDF:', gs_SVC.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Gaussian, CVEC:', gauss_gs.best_score_)
print('Testing Score, Gaussian, CVEC:', gauss_gs.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, Gaussian, TFIDF:', gauss_gs_tfidf.best_score_)
print('Testing Score, Gaussian, TFIDF:', gauss_gs_tfidf.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, ADABoost, CVEC:', gs_ADA_cv.best_score_)
print('Testing Score, ADABoost, CVEC:', gs_ADA_cv.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, ADABoost, TFIDF:', gs_ADA.best_score_)
print('Testing Score, ADABoost, TFIDF:', gs_ADA.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, GradientBoost, CVEC:', gs_gb_cv.best_score_)
print('Testing Score, GradientBoost, CVEC:', gs_gb_cv.score(X_test, y_test))
print()
print('='*100)
print()
print('Training Best Score, GradientBoost, TFIDF:', gs_gb.best_score_)
print('Testing Score, GradientBoost, TFIDF:', gs_gb.score(X_test, y_test))

The best model out of the models used seems to be the SVC using the TFIDFVectorizer

**Training Score**, SVC, TFIDF: 0.7728775156604025
<br>**Testing Score**, Extra SVC, TFIDF: 0.8333333333333334

print('Training Best Score, SVC, TFIDF:', gs_SVC.best_score_)
print('Testing Score, Extra SVC, TFIDF:', gs_SVC.score(X_test, y_test))
print()

# SVC Removing Stopwords

Trying out the best model(SVC) and checking how it performs with the lyrics that have the stopwords removed.

results = gs_SVC.fit(X_train, y_train)

results.score

SVC_pipe_test = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('SVC', SVC(random_state=42))
])

SVC_pipe_params_test = {
    'SVC__C': [1],
    'SVC__degree': [3],
    'SVC__kernel': ['linear'],
    'tfidf__max_df': [0.9],
    'tfidf__max_features': [3000],
    'tfidf__min_df': [1],
    'tfidf__ngram_range': [(1, 1)],
    'tfidf__stop_words': ['english']
}

gs_SVC_test = GridSearchCV(SVC_pipe_test,
                     SVC_pipe_params_test,
                     cv=5)

gs_SVC_test.fit(X_train, y_train)

gs_SVC_test.best_score_

gs_SVC_test.score(X_test, y_test)

The lyrics with the stopwords removed didn't perform as well as the lyrics with the stopwords left in. No need to try it on more models. I am going forward with the SVC model and TFIDFVectorizer.

## Grabbing the coefficients of the words

I am interested to see what words gave the greatest weight into predicting which songbook it belonged to. Also interested to see how differently the songs are weighted with the stopwords left in versus them being removed. 

songs['predictions_with_stopwords'] = gs_SVC_test.predict(songs['clean_lyrics_with_stopwords'])

# Grabbing the coef predictions with the stopwords removed
predictions_stopwords = pd.DataFrame(gs_SVC_test.best_estimator_.steps[1][1].coef_.toarray(),
             columns=gs_SVC_test.best_estimator_.steps[0][1].get_feature_names()).T.sort_values(0)

predictions_stopwords = predictions_stopwords.rename(columns={0: 'coefs_with_stopword'})

predictions_stopwords.head(20)

results.best_estimator_.steps[1][1].coef_.toarray()

# Coefs with stopwords left in
predictions = pd.DataFrame(results.best_estimator_.steps[1][1].coef_.toarray(),
             columns=results.best_estimator_.steps[0][1].get_feature_names()).T.sort_values(0)

predictions = predictions.rename(columns={0: 'coefs'})

# Combining into one dataframe for both coefs
# Then creating a CSV
coef_df = pd.concat([predictions, predictions_stopwords], axis=1)

coef_df.to_csv('datasets/coefs_with_and_without_stopwords.csv')

songs['predictions'] = gs_SVC.predict(songs['clean_lyrics_with_stopwords'])

songs.to_csv('datasets/predictions.csv', index=False)

# Pickles!

Using a pickle so I can take my model to a new notebook as to not overcrowd this one. 

import pickle

This was learned from the pickle local lesson in week 8.

with open('SVC.pkl', 'wb') as SCV_out:
    pickle.dump(gs_SVC, SCV_out)
