==> 06_Predictions_Data.ipynb <==
This notebook is just me looking at all the information I can about what my model got wrong.

import pandas as pd

songs = pd.read_csv('datasets/X_test_predictions.csv')

songs.drop(columns=['lyrics'], inplace=True)

# Looking at the values in test set

songs.shape

songs['disney_renaissance'].value_counts()

songs['american_songbook'].value_counts(normalize=True)

songs['american_songbook'].value_counts()

songs['predictions'].value_counts(normalize=True)

false_predictions = songs.loc[~(songs['predictions'] == songs['american_songbook']), :]

false_predictions['disney_renaissance'].value_counts(normalize=True)

false_predictions['american_songbook'].value_counts(normalize=True)

false_predictions['american_songbook'].value_counts()

false_predictions['predictions'].value_counts(normalize=True)

false_predictions.shape

songs.shape

disney_false = false_predictions.loc[false_predictions['american_songbook'] == 0, :]

disney_false['disney_renaissance'].value_counts(normalize=True)

Disney Renaissance makes up 38.7% of the Disney songs, but 50% of the false predictions

# Looking at incorrect predictions by songwriter

disney_false['song_writer'].value_counts()

false_predictions['song_writer'].value_counts()

# Looking at incorrect lyrics

# Mulan Make a Man out of You/
false_predictions.iloc[2]['clean_lyrics']

# The Emperor's New Groove My Funny Friend and Me
false_predictions.iloc[0]['clean_lyrics']

# George Gershwin Life Is Just A Bowl Of Cherries Lyrics
false_predictions.iloc[4]['clean_lyrics']

# Looking at the Coefficients

coefs = pd.read_csv('datasets/coefs_with_and_without_stopwords.csv')

coefs = coefs.rename(columns={'Unnamed: 0': 'words'})

coefs.head()

coefs['coef_differences'] = coefs['coefs'] - coefs['coefs_with_stopword']

coefs.head(20)

coefs.sort_values('coefs').head(50)

coefs.sort_values('coefs', ascending=False).head(50)

coefs.sort_values('coefs_with_stopword', ascending=False).head(50)

coefs.sort_values('coefs_with_stopword').head(50)

coefs.sort_values('coef_differences', ascending=False).head(50)

coefs.sort_values('coef_differences').head(50)
